Check if stream_copy_to_stream fails.

+++ src/Media/Process/Adapter/FfmpegShell.php
@@ -245,7 +245,9 @@ class FfmpegShell extends \mm\Media\Process\Adapter {
 
 		$target = fopen($targetTemp, 'r');
 		$buffer = fopen('php://temp', 'w+');
-		stream_copy_to_stream($target, $buffer);
+		if (!stream_copy_to_stream($target, $buffer)) {
+			throw new Exception('Failed to copy stream while processing ffmpeg source.');
+		}
 
 		fclose($target);
 		unlink($targetTemp);
